db라우터 작업
