+++ src/main.ts
@@ -9,6 +9,8 @@ var MySQLStore = require('express-mysql-session')(session);
 
 
 const cors = require('cors');
+const router = require('./routes/router');
+
 
 const multer = require('multer'); // express에 multer모듈 적용 (for 파일업로드)
 const upload = multer({
@@ -26,24 +28,24 @@ const upload = multer({
 
 
 
-const conn  = mysql.createConnection({
-    host     : '127.0.0.1',        // host명
-    user     : 'root',            // user명
-    password : 'Root!0000',            // 해당 user의 password
-    database : 'codemoa',            // DB명
-    port : '3306'                // mysql 접속 port
-});
+// const conn  = mysql.createConnection({
+//     host     : '127.0.0.1',        // host명
+//     user     : 'root',            // user명
+//     password : 'Root!0000',            // 해당 user의 password
+//     database : 'codemoa',            // DB명
+//     port : '3306'                // mysql 접속 port
+// });
 
 
  
 // Connect To database (DB에 직접 연결하는 )
-conn.connect(function(error:any) {
-    if (error) {
-        console.error('error connecting: ' + error.stack);
-        return;
-    }
-    console.log('connected as id ' + conn.threadId);
-});
+// conn.connect(function(error:any) {
+//     if (error) {
+//         console.error('error connecting: ' + error.stack);
+//         return;
+//     }
+//     console.log('connected as id ' + conn.threadId);
+// });
 
 export class Server {
 
@@ -71,7 +73,6 @@ this.config();
 
 
 this.routes();
-
 
 
 }
@@ -116,224 +117,224 @@ console.log(`Listening at http://localhost:${this.PORT}/`);
 
 
 
-  this.app.get('/user', function(req, res, next){
-    var sql = 'select * from question';
-    conn.query(sql, function(err:any,data:any){
-           if(err) console.log(err);
+//   this.app.get('/user', function(req, res, next){
+//     var sql = 'select * from question';
+//     conn.query(sql, function(err:any,data:any){
+//            if(err) console.log(err);
 
-        var string = JSON.stringify(data)         
-        var json = JSON.parse(string);
-        console.log('json: ', json);
+//         var string = JSON.stringify(data)         
+//         var json = JSON.parse(string);
+//         console.log('json: ', json);
             
-        res.send(json);
-    });
-  });
+//         res.send(json);
+//     });
+//   });
 
-  this.app.get('/myQuestion', function(req: any, res, next){
+//   this.app.get('/myQuestion', function(req: any, res, next){
 
-    var email = req.query.m_email;
+//     var email = req.query.m_email;
 
-    console.log("reqemail", email);
+//     console.log("reqemail", email);
     
 
-    var sql = 'select * from question where m_email = ?';
-    conn.query(sql,[email],function(err:any,data:any){
-           if(err) console.log(err);
+//     var sql = 'select * from question where m_email = ?';
+//     conn.query(sql,[email],function(err:any,data:any){
+//            if(err) console.log(err);
 
-        var string = JSON.stringify(data)         
-        var json = JSON.parse(string);
-        console.log('json: ', json);
+//         var string = JSON.stringify(data)         
+//         var json = JSON.parse(string);
+//         console.log('json: ', json);
             
-        res.send(json);
-    });
-  });
+//         res.send(json);
+//     });
+//   });
 
-  this.app.get('/user/question', function(req:any, res){
+//   this.app.get('/user/question', function(req:any, res){
     
-    var q_id = req.query.q_id;
+//     var q_id = req.query.q_id;
 
-    var sql = 'select * from question where q_id = ?';
-    conn.query(sql, [q_id], function(err:any,data:any){
-        if(err) console.log(err);
+//     var sql = 'select * from question where q_id = ?';
+//     conn.query(sql, [q_id], function(err:any,data:any){
+//         if(err) console.log(err);
 
-        var string = JSON.stringify(data)         
-        var json = JSON.parse(string);
-        console.log('question: ', json);
+//         var string = JSON.stringify(data)         
+//         var json = JSON.parse(string);
+//         console.log('question: ', json);
             
-        res.send(json);
-    });
-  });
+//         res.send(json);
+//     });
+//   });
 
-  this.app.post('/login', function(req:any, res){
-     req.body = Object.keys(req.body);
-     req.body = JSON.parse(req.body[0]);
-     //var array  = Object.keys(req.body);
+//   this.app.post('/login', function(req:any, res){
+//      req.body = Object.keys(req.body);
+//      req.body = JSON.parse(req.body[0]);
+//      //var array  = Object.keys(req.body);
 
-     var email = req.body.loginid; 
-     var pw = req.body.loginpw;
+//      var email = req.body.loginid; 
+//      var pw = req.body.loginpw;
 
-     console.log(req);
-     console.log("main", req.body.loginid);
+//      console.log(req);
+//      console.log("main", req.body.loginid);
      
 
-     var sql = 'select * from member where m_email = ?';
-     let logS = true;
+//      var sql = 'select * from member where m_email = ?';
+//      let logS = true;
      
 
-     conn.query(sql, [email], function(err:any, data:any){
-        if(err)
-            console.log("로그인 에러남", err);
+//      conn.query(sql, [email], function(err:any, data:any){
+//         if(err)
+//             console.log("로그인 에러남", err);
         
-        var string = JSON.stringify(data)         
-        var json = JSON.parse(string);
-        console.log('데이터값: ', json);
+//         var string = JSON.stringify(data)         
+//         var json = JSON.parse(string);
+//         console.log('데이터값: ', json);
 
-        if(!json[0]){
-            logS = false;
-            //res.render('user', { error: '로그인실패'});
-            res.send(logS);
+//         if(!json[0]){
+//             logS = false;
+//             //res.render('user', { error: '로그인실패'});
+//             res.send(logS);
         
-        }else if(json[0].m_password != pw){
-            logS = false;
-            res.send(logS);
-            //res.redirect('/');
-            console.log('로그인실패');
+//         }else if(json[0].m_password != pw){
+//             logS = false;
+//             res.send(logS);
+//             //res.redirect('/');
+//             console.log('로그인실패');
             
-        }else if(json[0].m_password == pw){
-            // req.session.email = email;
-            // console.log('email저장됐니?', req.session.email);
+//         }else if(json[0].m_password == pw){
+//             // req.session.email = email;
+//             // console.log('email저장됐니?', req.session.email);
            
-            logS = true;
+//             logS = true;
 
             
 
-            console.log('로그인성공');
-            res.send(json[0]);
-            //res.redirect('/');
-        }
+//             console.log('로그인성공');
+//             res.send(json[0]);
+//             //res.redirect('/');
+//         }
 
-     });      
-  });
+//      });      
+//   });
 
-  this.app.post('/questInput', function(req:any, res){
-    req.body = Object.keys(req.body);
-    req.body = JSON.parse(req.body[0]);
+//   this.app.post('/questInput', function(req:any, res){
+//     req.body = Object.keys(req.body);
+//     req.body = JSON.parse(req.body[0]);
 
-    var name = req.body.name; 
-    var email = req.body.email;
-    var title = req.body.title;
-    var content = req.body.content;
-    var category = req.body.category;
+//     var name = req.body.name; 
+//     var email = req.body.email;
+//     var title = req.body.title;
+//     var content = req.body.content;
+//     var category = req.body.category;
 
-    console.log(req);
+//     console.log(req);
 
-    // content = content.replace(/\n/g, "<br>");
-    // content = content.replaceAll("<br>", "\r\n");
+//     // content = content.replace(/\n/g, "<br>");
+//     // content = content.replaceAll("<br>", "\r\n");
     
 
-    var sql = 'insert into question values(0,?,?,?,?,0,?)';
-    let logS = true;
+//     var sql = 'insert into question values(0,?,?,?,?,0,?)';
+//     let logS = true;
     
 
-    conn.query(sql, [email, name, title, content, category], function(err:any, data:any){
-       if(err){
-           console.log("에러남", err);
-           res.send(false); 
-       }else{
-           res.send(true);
-       }
+//     conn.query(sql, [email, name, title, content, category], function(err:any, data:any){
+//        if(err){
+//            console.log("에러남", err);
+//            res.send(false); 
+//        }else{
+//            res.send(true);
+//        }
 
 
-    });      
- });
+//     });      
+//  });
  
- this.app.post('/register', function(req:any, res){
+//  this.app.post('/register', function(req:any, res){
 
-    req.body = Object.keys(req.body);
-    req.body = JSON.parse(req.body[0]);
+//     req.body = Object.keys(req.body);
+//     req.body = JSON.parse(req.body[0]);
 
-    var name = req.body.m_name; 
-    var email = req.body.m_email;
-    var password = req.body.m_password;
+//     var name = req.body.m_name; 
+//     var email = req.body.m_email;
+//     var password = req.body.m_password;
 
-    console.log(req);
-    console.log("main", req.body.m_name);
+//     console.log(req);
+//     console.log("main", req.body.m_name);
     
 
-    var sql = 'insert into member values(0,?,?,?)';
-    let logS = true;
+//     var sql = 'insert into member values(0,?,?,?)';
+//     let logS = true;
     
 
-    conn.query(sql, [email, password, name], function(err:any, data:any){
-       if(err){
-           console.log("에러남", err);
-           res.send(false); 
-       }else{
-           res.send(true);
-       }
+//     conn.query(sql, [email, password, name], function(err:any, data:any){
+//        if(err){
+//            console.log("에러남", err);
+//            res.send(false); 
+//        }else{
+//            res.send(true);
+//        }
 
 
-    });      
- });
+//     });      
+//  });
  
- this.app.post('/awnserInput', function(req:any, res){
-    req.body = Object.keys(req.body);
-    req.body = JSON.parse(req.body[0]);
+//  this.app.post('/awnserInput', function(req:any, res){
+//     req.body = Object.keys(req.body);
+//     req.body = JSON.parse(req.body[0]);
 
-    var q_id = req.body.q_id; 
-    var q_name = req.body.q_name;
-    var a_name = req.body.a_name;
-    var content = req.body.content
+//     var q_id = req.body.q_id; 
+//     var q_name = req.body.q_name;
+//     var a_name = req.body.a_name;
+//     var content = req.body.content
 
-    console.log(req);
-    console.log("main", req.body.loginid);
+//     console.log(req);
+//     console.log("main", req.body.loginid);
     
 
-    var sql = 'insert into awnser values(0,?,?,?,?,0)';
-    var sql2 = 'update question set awn_count = awn_count + 1 where q_id = '+q_id;
-    let logS = true;
+//     var sql = 'insert into awnser values(0,?,?,?,?,0)';
+//     var sql2 = 'update question set awn_count = awn_count + 1 where q_id = '+q_id;
+//     let logS = true;
     
 
-    conn.query(sql, [q_id, q_name, a_name, content], function(err:any, data:any){
-       if(err){
-           console.log("에러남", err);
-           res.send(false); 
-       }else{
+//     conn.query(sql, [q_id, q_name, a_name, content], function(err:any, data:any){
+//        if(err){
+//            console.log("에러남", err);
+//            res.send(false); 
+//        }else{
 
-        conn.query(sql2, function(){
-            if(err)
-                console.log("조회수 에러");    
-        });
+//         conn.query(sql2, function(){
+//             if(err)
+//                 console.log("조회수 에러");    
+//         });
 
-           res.send(true);
-       }
+//            res.send(true);
+//        }
 
 
-    });      
- });
+//     });      
+//  });
 
- this.app.post('/awnserList', function(req:any, res){
-    req.body = Object.keys(req.body);
-    req.body = JSON.parse(req.body[0]);
+//  this.app.post('/awnserList', function(req:any, res){
+//     req.body = Object.keys(req.body);
+//     req.body = JSON.parse(req.body[0]);
 
-    var q_id = req.body.q_id;
+//     var q_id = req.body.q_id;
 
-    //console.log(req);
-    console.log("main", req.body);
+//     //console.log(req);
+//     console.log("main", req.body);
     
 
-    var sql = 'select * from awnser where q_id = ?';
+//     var sql = 'select * from awnser where q_id = ?';
 
-    conn.query(sql, [q_id], function(err:any, data:any){
-       if(err)
-           console.log("에러남", err);
+//     conn.query(sql, [q_id], function(err:any, data:any){
+//        if(err)
+//            console.log("에러남", err);
 
-        var string = JSON.stringify(data)         
-        var json = JSON.parse(string);
+//         var string = JSON.stringify(data)         
+//         var json = JSON.parse(string);
 
-        res.send(json);
-    });      
- });
+//         res.send(json);
+//     });      
+//  });
 
 //  this.app.post('/upload', upload, function(req:any, res){
 //     res.send('Upload'+req.file);
@@ -365,6 +366,7 @@ routes() {
 
 
 
+    this.app.use('/', router);
 }
 
 public static bootstrap(): Server {
